修复 可能的没有 human 消息的错误

+++ internal/base/conf/conf.go
@@ -92,8 +92,8 @@ type S3 struct {
 }
 
 type LLM struct {
-	MaxTokens                  int   `yaml:"max_tokens" mapstructure:"max_tokens"`
-	ContextOptimizeActiveCount int64 `yaml:"context_optimize_active_count" mapstructure:"context_optimize_active_count"`
+	MaxTokens                  int `yaml:"max_tokens" mapstructure:"max_tokens"`
+	ContextOptimizeActiveCount int `yaml:"context_optimize_active_count" mapstructure:"context_optimize_active_count"`
 	//Temperature float64 `yaml:"temperature"`
 	//TopP float64 `yaml:"top_p"`
 	//N int `yaml:"n"`
